+++ HLSpriteKit/SKLabelNode+HLLabelNodeAdditions.m
@@ -77,7 +77,7 @@ forHLVerticalAlignmentMode:(HLLabelNodeVerticalAlignmentMode)hlVerticalAlignment
     }
 
     default:
-      [NSException raise:@"HLLabelNodeUnknownVerticalAlignmentMode" format:@"Unknown vertical alignment mode %d.", hlVerticalAlignmentMode];
+      [NSException raise:@"HLLabelNodeUnknownVerticalAlignmentMode" format:@"Unknown vertical alignment mode %ld.", (long)hlVerticalAlignmentMode];
   }
 }
